packages/okra-pg: name PostgresTree
